feat(ios): complete iOS deployment configuration and build setup
- Updated Info.plist with proper permissions and app configuration
- Created Podfile with Firebase and iOS dependencies
- Configured entitlements for development and production
- Added build scripts for archive creation and IPA export
- Set up fastlane with release, beta, and development lanes
- Created comprehensive CI/CD pipeline in codemagic.yaml
- Added complete iOS deployment documentation and guides
- Ready for macOS build and App Store submission

+++ lib/services/ab_test_service.dart
@@ -40,7 +40,7 @@ class ABTestService {
       }
 
       final totalTraffic =
-          variants.fold(0, (sum, variant) => sum + variant.trafficPercentage);
+          variants.fold(0.0, (sum, variant) => sum + variant.trafficPercentage);
       if (totalTraffic > 100.0) {
         throw Exception('Общий трафик вариантов не может превышать 100%');
       }
@@ -264,7 +264,7 @@ class ABTestService {
 
     var cumulativePercentage = 0;
     for (final variant in test.variants) {
-      cumulativePercentage += variant.trafficPercentage / 100.0;
+      cumulativePercentage += (variant.trafficPercentage / 100.0).toInt();
       if (randomValue <= cumulativePercentage) {
         return variant;
       }
